upload bioimage to different folder according to the profile type

+++ src/apps/copo_dtol_submission/utils/Dtol_Bioimage_Submission.py
@@ -11,22 +11,35 @@ from common.utils.helpers import get_env, get_datetime, notify_frontend
 
 BIOIMAGE_SERVER = get_env("BIOIMAGE_SERVER")  # "bsaspera_w@hx-fasp-1.ebi.ac.uk"
 BIOIMAGE_UPLOAD_PATH = get_env("BIOIMAGE_PATH")  # "/.beta/91/31c15a-f0a0-4847-ab09-4135cefc03bd-a31912"
-BIOIMAGE_LOCAL_ARCHIVE_PATH = f"{get_env('MEDIA_PATH')}sample_images/archive/sent"
+BIOIMAGE_LOCAL_ARCHIVE_PARENT_PATH = f"{get_env('MEDIA_PATH')}sample_images/archive"
+BIOIMAGE_LOCAL_ARCHIVE_PATH = {"asg_specimen" : f"{BIOIMAGE_LOCAL_ARCHIVE_PARENT_PATH}/ASG",
+                             "dtol_specimen" : f"{BIOIMAGE_LOCAL_ARCHIVE_PARENT_PATH}/DTOL",
+                             "erga_specimen" : f"{BIOIMAGE_LOCAL_ARCHIVE_PARENT_PATH}/ERGA",
+                             "copo_specimen" : f"{BIOIMAGE_LOCAL_ARCHIVE_PARENT_PATH}/COPO"}
+
 BIOIMAGE_PATH = f"{settings.MEDIA_ROOT}sample_images"
 BIOIMAGE_ARCHIVE = f"{BIOIMAGE_PATH}/archive"
-BIOIMAGE_SENT = f"{BIOIMAGE_PATH}/COPO"
 BIOIMAGE_THUMBNAIL = f"{BIOIMAGE_PATH}/thumbnail"
 
+BIOIMAGE_SENT = {"asg_specimen" : f"{BIOIMAGE_PATH}/ASG",
+                 "dtol_specimen" : f"{BIOIMAGE_PATH}/DTOL",
+                 "erga_specimen" : f"{BIOIMAGE_PATH}/ERGA",
+                 "copo_specimen" : f"{BIOIMAGE_PATH}/COPO"}
+
 ASPERA_PATH = get_env("ASPERA_PATH")  #"/root/.aspera/cli"
-BIOIMAGE_ASPERA_CMD = f"{ASPERA_PATH}/bin/ascp -P33001 -l700M --move-after-transfer  {BIOIMAGE_ARCHIVE} -i {ASPERA_PATH}/etc/asperaweb_id_dsa.openssh -d {BIOIMAGE_SENT} {BIOIMAGE_SERVER}:{BIOIMAGE_UPLOAD_PATH}"
+BIOIMAGE_ASPERA_CMD = f"{ASPERA_PATH}/bin/ascp -P33001 -l700M --move-after-transfer  {BIOIMAGE_ARCHIVE} -i {ASPERA_PATH}/etc/asperaweb_id_dsa.openssh -d {' '.join(BIOIMAGE_SENT.values())} {BIOIMAGE_SERVER}:{BIOIMAGE_UPLOAD_PATH}"
 
 
 def housekeeping_bioimage_archive():
     housekeep_timestamp = datetime.timestamp(datetime.now() + timedelta(days=-30))
-    with os.scandir(BIOIMAGE_ARCHIVE+"/COPO") as ls:
-        for file in ls:
-            if os.path.getctime(file) < housekeep_timestamp:
-                os.remove(file)
+    for type in 'ASG', 'DTOL', 'ERGA', 'COPO':
+        try:
+            with os.scandir(BIOIMAGE_ARCHIVE + "/" + type) as ls:
+                for file in ls:
+                    if file.is_file() and os.path.getctime(file) < housekeep_timestamp:
+                        os.remove(file)
+        except FileNotFoundError as e:
+            pass
 
 def process_bioimage_pending_submission():
     # submit images
@@ -38,7 +51,8 @@ def process_bioimage_pending_submission():
     #imagePath = Path(settings.MEDIA_ROOT) / "sample_images"
     #sentPath = imagePath / "sent"
 
-    Path(BIOIMAGE_SENT).mkdir(parents=True, exist_ok=True)
+    for path in BIOIMAGE_SENT.values():
+        Path(path).mkdir(parents=True, exist_ok=True)
 
     if not submissions:
         return
@@ -52,6 +66,7 @@ def process_bioimage_pending_submission():
     sources = Source().get_sourcemap_by_specimens(specimen_ids)
     for specimenId in specimen_ids:
         source = sources[specimenId]
+        type = source.get("sample_type", "copo_specimen")
         seqno = 0
         #if "last_bioimage_submitted" in source and  source["last_bioimage_submitted"]:
         #    lastSubImageDt[specimenId] = source["last_bioimage_submitted"]
@@ -65,20 +80,27 @@ def process_bioimage_pending_submission():
                         ##if specimenId not in lastSubImageDt or os.path.getctime(imageFile) > datetime.timestamp(
                         ##        lastSubImageDt[specimenId]):
                         newname = source["biosampleAccession"] + "_" + str(seqno+1) + os.path.splitext(imageFile)[1]
-                        DataFile().update_bioimage_name(imageFile.name, newname, f"{BIOIMAGE_LOCAL_ARCHIVE_PATH}/{newname}")
-                        os.rename(imageFile.path, Path(BIOIMAGE_SENT) / newname)
+                        DataFile().update_bioimage_name(imageFile.name, newname, f"{BIOIMAGE_LOCAL_ARCHIVE_PATH[type]}/{newname}")
+                        os.rename(imageFile.path, Path( BIOIMAGE_SENT[type]) / newname) 
                         try:
                             os.remove(f"{BIOIMAGE_THUMBNAIL}/{imageFile.name}")
                         except FileNotFoundError as e:
                             Logger().log(e, level=Loglvl.ERROR)
                         seqno = seqno + 1
                         print(imageFile.name + " " + newname)
+        
         finally:
             Source().add_fields({"last_bioimage_submitted": now,
                                      "bioimage_archive_seq_no": seqno,
                                      "date_modified": now}, source["_id"])
 
-    if len(os.listdir(BIOIMAGE_SENT)) > 0:
+    is_sent = False
+    for path in BIOIMAGE_SENT.values():
+        if len(os.listdir(path)) > 0:
+            is_sent = True
+            break
+
+    if is_sent:
         try:
             Logger().log(BIOIMAGE_ASPERA_CMD, level=Loglvl.DEBUG)
             output = subprocess.check_output(BIOIMAGE_ASPERA_CMD, shell=True)
@@ -97,10 +119,6 @@ def process_bioimage_pending_submission():
         notify_frontend(data={"profile_id": sub["profile_id"]}, msg="Bioimage not submitted", action="info",
                         html_id="dtol_sample_info")
 
-
-
     Submission().get_collection_handle().update_many(
         {"_id" : {"$in": sub_ids}}, {"$set": {"dtol_specimen": [], "dtol_status":"complete", "date_modified": now}}
     )
-
-
